% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilibiliDownload.R
\name{bilibiliConcat}
\alias{bilibiliConcat}
\title{Concatenates videos from Bilibili}
\usage{
bilibiliConcat(fileFormat = "flv", path = "./", removeSource = FALSE)
}
\arguments{
\item{fileFormat}{the format of videos to concatenate, the default is "flv".}

\item{path}{the path of videos located, the default is current directory.}

\item{removeSource}{logic, remove source videos if TURE.}
}
\description{
Concatenates videos from Bilibili
}
