% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilibiliDownload.R
\name{bilibiliDownload}
\alias{bilibiliDownload}
\title{Download (a list of ) Bilibili videos by providing a url seed.}
\usage{
bilibiliDownload(
  urlSeed = urlSeed,
  quality = c(p720p = 64, p360p = 16, p480p = 32, p1080p = 80)[1],
  path = "./bilibili_video/",
  removeSource = FALSE,
  sleepTime = 5,
  method,
  ...
)
}
\arguments{
\item{urlSeed}{a url string.}

\item{quality}{the quality of the video(s), 64 (default) is 720p, 16 is 360p, 32 is 480p, 80 is 1080p.}

\item{path}{the path to save the list of videos.}

\item{removeSource}{logic, whether to remove the source video after concatenate the source videos.}
}
\description{
Bilibili API collection: https://www.bilibili.com/read/cv3430609/ \cr
The API to get AV ID by BV ID:
https://api.bilibili.com/x/web-interface/view?bvid=BVID
}
\examples{
{
\dontrun{
# Download a single video if there is "/?p=num" in the urlSeed.
urlSeed = "https://www.bilibili.com/video/BV1kb411x7KZ?p=1"
bilibiliDownload(urlSeed)

# Download a list of videos if there is no "/]?p=num" in the urlSeed.
urlSeed = "https://www.bilibili.com/video/BV1kb411x7KZ"
bilibiliDownload(urlSeed = urlSeed,
                 quality = 32,
                 path = "./小学二年级数学上册【杨老师54讲】",
                 removeSource = TRUE)

}
}
}
