% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio2mp4.R
\name{video2mp4}
\alias{video2mp4}
\title{Transform other video format to .mp4}
\usage{
video2mp4(fileFormat = "webm", path, removeSource = FALSE)
}
\arguments{
\item{fileFormat}{character, video file format, without "." in front.
The default is "webm".}

\item{path}{character, the path where the files located.}

\item{removeSource}{logic, removing the original video file after
transformation.}
}
\description{
Transform other video format to .mp4
}
