% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoListDownloadByTable.R
\name{videoListDownloadByTable}
\alias{videoListDownloadByTable}
\title{Download YouTube videos/audios by providing a table of video/audio list.}
\usage{
videoListDownloadByTable(
  fileList,
  fileTable = NULL,
  path,
  bothVideoAudio = TRUE,
  priority = c("mp4", "best", "audio only"),
  id = NULL,
  sleepTime = 10
)
}
\arguments{
\item{fileList}{the file name of video/audio list table. This table
is a csv file generated by \code{\link{videoListTable}} function.
The columns of this table are 'rowNumber', 'Orders',
'VideoTitles', 'fileName', and 'URL', in which 'URL' is most
important information. If fileTable is provided, then this argument is
not used.}

\item{fileTable}{the file table returned by \code{\link{videoListTable}}
function. If fileTable is provided, then fileList argument is
not used.}

\item{path}{the path to save video/audio files.}

\item{bothVideoAudio}{logic, whether download both audio and video,
it should be FALSE if only downloading audio or video. The default
is \code{TRUE}.}

\item{priority}{the file format to download. The option can be
any one (or combination) of "mp4", "best", "audio only", "mp3",
"webm", the default
is c("mp4", "best", "audio only"), which means the downloader will
look for mp4 file first, if this file does not exist, then the
downloader will look for the best file marked by YouTube, then
look for only audio file.}

\item{id}{numeric or \code{NULL}, the index of rows of the `fileList`
table to download. The default is \code{NULL}.}

\item{sleepTime}{numeric, the time to pause to prevent YouTube blocking.
The default is 10 second}
}
\description{
Download YouTube videos/audios by providing a table of video/audio list.
}
\examples{
{
\dontrun{
file = "fileNameOrders.csv"
folderName = "./fileByTable"
#### download video by fileNameOrders.csv table
newFolder = videoListDownloadByTable(fileList = file,
                                     folderName = folderName,
                                     bothVideoAudio = TRUE,
                                     priority = c("best"),
                                     sleepTime = 10,
                                     id = c(1, 2, 5, 9))
}
}
}
