% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio2mp3.R
\name{audio2mp3}
\alias{audio2mp3}
\title{Transform other audio format to .mp3}
\usage{
audio2mp3(fileFormat = "webm", path, removeSource = FALSE)
}
\arguments{
\item{fileFormat}{character, audio file format, without "." in front.
The default is "webm".}

\item{path}{character, the path where the files located.}

\item{removeSource}{logic, removing the original audio file after
transformation.}
}
\description{
Transform other audio format to .mp3
}
