% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilibiliDownload.R
\name{bilibiliDownload}
\alias{bilibiliDownload}
\title{Download (a list of ) Bilibili videos by providing a url seed.}
\usage{

  bilibiliDownload(urlSeed = "https://www.bilibili.com/video/av30300809/?p=40",
  quality = c(p720p = 64, p360p = 16, p480p = 32, p1080p = 80)[1],
  path = "./bilibili_video/", removeSource = FALSE)
}
\arguments{
\item{urlSeed}{a url string.}

\item{quality}{the quality of the video(s), 64 (default) is 720p, 16 is 360p, 32 is 480p, 80 is 1080p.}

\item{path}{the path to save the list of videos.}

\item{removeSource}{logic, whether to remove the source video after concatenate the source videos.}
}
\description{
Download (a list of ) Bilibili videos by providing a url seed.
}
\examples{
{
\dontrun{
# Download a single video if there is "/?p=num" in the urlSeed.
bilibiliDownload("https://www.bilibili.com/video/av30300809/?p=40")

# Download a list of videos if there is no "/?p=num" in the urlSeed.
bilibiliDownload(urlSeed = "https://www.bilibili.com/video/av30300809",
                 quality = 32,
                 path = "./python数据分析与机器学习",
                 removeSource = TRUE)

}
}
}
