% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilibiliDownload.R
\name{bilibiliListTable}
\alias{bilibiliListTable}
\title{Downloading by an URL seed}
\usage{
bilibiliListTable(
  urlSeed,
  path = "./bilibili_video/",
  saveFileList = TRUE,
  addOrder = FALSE
)
}
\arguments{
\item{path}{the path to save video/audio files.}

\item{saveFileList}{logic, whether save the file table.}

\item{addOrder}{logic, whether add a formated number to the videoTitles, default is FALSE.}

\item{fileList}{the file name of video/audio list table. This table
is a csv file generated by \code{\link{videoListDownload}} function.
The columns of this table are 'rowNames', 'data.index',
'data.video.title', 'fileName', and 'URL', in which 'URL' is most
important information.}
}
\description{
Download Bilibili videos/audios by providing an URL seed,
which is any URL in a video list.
}
\examples{
{
\dontrun{
urlSeed = "https://www.bilibili.com/video/BV1kb411x7KZ"
bilibiliListTable(urlSeed)

bilibiliListTable(urlSeed = urlSeed,
                  path = "./小学二年级数学上册【杨老师54讲】",
                  addOrder = TRUE)

}
}
}
