% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youTubeDownload.R
\name{youTubeDownload}
\alias{youTubeDownload}
\title{Download one single video/audio}
\usage{
youTubeDownload(url, path = getwd(), saveFile = "videoFromYoutube",
  priority = c("mp4", "best", "audio only"), bothVideoAudio = TRUE)
}
\arguments{
\item{url}{the YouTube url to download the video/audio file.}

\item{path}{the path to save downloaded file. The default is
the current directory.}

\item{saveFile}{the downloaded file name without file extension.
The default is "videoFromYoutube".}

\item{priority}{the file format to download. The option can be
any one (or combination) of "mp4", "best", "audio only", "mp3",
"webm", the default
is c("mp4", "best", "audio only"), which means the downloader will
look for mp4 file first, if this file does not exist, then the
downloader will look for the best file marked by YouTube, then
look for only audio file.}

\item{bothVideoAudio}{logic, whether download both audio and video,
it should be FALSE if only downloading audio or video. The default
is \code{TRUE}.}
}
\description{
Download one single video/audio
}
\examples{
{
\dontrun{
# Downlad an audio
url0 = "https://www.youtube.com/watch?v=J9NQFACZYEU&list=PLMC9KNkIncKtPzgY-5rmhvj7fax8fdxoj&index=PLMC9KNkIncKtPzgY-5rmhvj7fax8fdxoj"
youTubeDownload(url0, path = "./music",
                saveFile = "videoFromYoutube",
                priority = "audio only",
                bothVideoAudio = FALSE)

# Downlad a video
url0 = "https://www.youtube.com/watch?v=pJON0-e_I3o&t=889s"
youTubeDownload(url0, path = "./OneVideo",
                saveFile = "videoFile",
                priority = c("best", "mp4"))
}
}
}
